rename src to q_kernel_ops
