+++ q_kernel_ops/controllers/kernel.py
@@ -14,9 +14,9 @@ import copy
 import math
 from pympler import asizeof
 
-from src.circuits import circuit_2, circuit_5, circuit_10, circuit_18, kernel_circuit
-from src.data import kernel_metadata, kernel_telemetry
-from src.runtime import Runtime
+from q_kernel_ops.circuits import circuit_2, circuit_5, circuit_10, circuit_18, kernel_circuit
+from q_kernel_ops.data import kernel_metadata, kernel_telemetry
+from q_kernel_ops.runtime import Runtime
 
 
 class Kernel:
