+++ tox.ini
@@ -21,7 +21,7 @@ commands =
 envdir = .tox/lint
 sitepackages = False
 commands =
-  pylint -rn src tests
+  pylint -rn q_kernel_ops tests
 
 [testenv:coverage]
 basepython = python3
@@ -31,7 +31,7 @@ commands =
 
 [testenv:black]
 envdir = .tox/lint
-commands = black {posargs} src tests --check
+commands = black {posargs} q_kernel_ops tests --check
 
 [testenv:json]
 allowlist_externals = /bin/bash
